pyproject.toml
[project]
name = "rpg-mcp-api"
version = "0.1.0"
description = "FastAPI SSE transport for RPG MCP Server"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-dotenv>=1.0.0",
    "mcp>=1.21.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
